changed: move Interface class out of DomainDecomposition
will be reused for intra-process connection representations
